+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Api/Controllers/AggregateTestNoIdReturnsController.cs
@@ -80,6 +80,7 @@ namespace CleanArchitecture.TestApplication.Api.Controllers
             [FromBody] UpdateAggregateTestNoIdReturnCommand command,
             CancellationToken cancellationToken = default)
         {
+            command.SetId(id);
             if (id != command.Id)
             {
                 return BadRequest();