+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Api/Controllers/ImplicitKeyAggrRootsController.cs
@@ -132,6 +132,7 @@ namespace CleanArchitecture.TestApplication.Api.Controllers
             [FromBody] UpdateImplicitKeyAggrRootCommand command,
             CancellationToken cancellationToken = default)
         {
+            command.SetId(id);
             if (id != command.Id)
             {
                 return BadRequest();
@@ -157,6 +158,7 @@ namespace CleanArchitecture.TestApplication.Api.Controllers
             [FromBody] UpdateImplicitKeyAggrRootImplicitKeyNestedCompositionCommand command,
             CancellationToken cancellationToken = default)
         {
+            command.SetId(id);
             if (implicitKeyAggrRootId != command.ImplicitKeyAggrRootId)
             {
                 return BadRequest();