+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Api/Controllers/WithCompositeKeysController.cs
@@ -4,13 +4,12 @@ using System.Net.Mime;
 using System.Threading;
 using System.Threading.Tasks;
 using CleanArchitecture.TestApplication.Api.Controllers.ResponseTypes;
-using CleanArchitecture.TestApplication.Application.AggregateRootLongs;
-using CleanArchitecture.TestApplication.Application.AggregateRootLongs.CreateAggregateRootLong;
-using CleanArchitecture.TestApplication.Application.AggregateRootLongs.DeleteAggregateRootLong;
-using CleanArchitecture.TestApplication.Application.AggregateRootLongs.GetAggregateRootLongById;
-using CleanArchitecture.TestApplication.Application.AggregateRootLongs.GetAggregateRootLongs;
-using CleanArchitecture.TestApplication.Application.AggregateRootLongs.UpdateAggregateRootLong;
-using CleanArchitecture.TestApplication.Application.Common.Pagination;
+using CleanArchitecture.TestApplication.Application.WithCompositeKeys;
+using CleanArchitecture.TestApplication.Application.WithCompositeKeys.CreateWithCompositeKey;
+using CleanArchitecture.TestApplication.Application.WithCompositeKeys.DeleteWithCompositeKey;
+using CleanArchitecture.TestApplication.Application.WithCompositeKeys.GetWithCompositeKeyById;
+using CleanArchitecture.TestApplication.Application.WithCompositeKeys.GetWithCompositeKeys;
+using CleanArchitecture.TestApplication.Application.WithCompositeKeys.UpdateWithCompositeKey;
 using Intent.RoslynWeaver.Attributes;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
@@ -25,11 +24,11 @@ using Microsoft.Extensions.DependencyInjection;
 namespace CleanArchitecture.TestApplication.Api.Controllers
 {
     [ApiController]
-    public class AggregateRootLongsController : ControllerBase
+    public class WithCompositeKeysController : ControllerBase
     {
         private readonly ISender _mediator;
 
-        public AggregateRootLongsController(ISender mediator)
+        public WithCompositeKeysController(ISender mediator)
         {
             _mediator = mediator ?? throw new ArgumentNullException(nameof(mediator));
         }
@@ -38,17 +37,17 @@ namespace CleanArchitecture.TestApplication.Api.Controllers
         /// </summary>
         /// <response code="201">Successfully created.</response>
         /// <response code="400">One or more validation errors have occurred.</response>
-        [HttpPost("api/aggregate-root-longs")]
+        [HttpPost("api/with-composite-key")]
         [Produces(MediaTypeNames.Application.Json)]
-        [ProducesResponseType(typeof(JsonResponse<long>), StatusCodes.Status201Created)]
+        [ProducesResponseType(typeof(JsonResponse<Guid>), StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult<JsonResponse<long>>> CreateAggregateRootLong(
-            [FromBody] CreateAggregateRootLongCommand command,
+        public async Task<ActionResult<JsonResponse<Guid>>> CreateWithCompositeKey(
+            [FromBody] CreateWithCompositeKeyCommand command,
             CancellationToken cancellationToken = default)
         {
             var result = await _mediator.Send(command, cancellationToken);
-            return CreatedAtAction(nameof(GetAggregateRootLongById), new { id = result }, new JsonResponse<long>(result));
+            return Created(string.Empty, new JsonResponse<Guid>(result));
         }
 
         /// <summary>
@@ -56,16 +55,17 @@ namespace CleanArchitecture.TestApplication.Api.Controllers
         /// <response code="200">Successfully deleted.</response>
         /// <response code="400">One or more validation errors have occurred.</response>
         /// <response code="404">One or more entities could not be found with the provided parameters.</response>
-        [HttpDelete("api/aggregate-root-longs/{id}")]
+        [HttpDelete("api/with-composite-key/{key1Id}/{key2Id}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult> DeleteAggregateRootLong(
-            [FromRoute] long id,
+        public async Task<ActionResult> DeleteWithCompositeKey(
+            [FromRoute] Guid key1Id,
+            [FromRoute] Guid key2Id,
             CancellationToken cancellationToken = default)
         {
-            await _mediator.Send(new DeleteAggregateRootLongCommand(id: id), cancellationToken);
+            await _mediator.Send(new DeleteWithCompositeKeyCommand(key1Id: key1Id, key2Id: key2Id), cancellationToken);
             return Ok();
         }
 
@@ -74,18 +74,23 @@ namespace CleanArchitecture.TestApplication.Api.Controllers
         /// <response code="204">Successfully updated.</response>
         /// <response code="400">One or more validation errors have occurred.</response>
         /// <response code="404">One or more entities could not be found with the provided parameters.</response>
-        [HttpPut("api/aggregate-root-longs/{id}")]
+        [HttpPut("api/with-composite-key/{key1Id}/{key2Id}")]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult> UpdateAggregateRootLong(
-            [FromRoute] long id,
-            [FromBody] UpdateAggregateRootLongCommand command,
+        public async Task<ActionResult> UpdateWithCompositeKey(
+            [FromRoute] Guid key1Id,
+            [FromRoute] Guid key2Id,
+            [FromBody] UpdateWithCompositeKeyCommand command,
             CancellationToken cancellationToken = default)
         {
-            command.SetId(id);
-            if (id != command.Id)
+            command.SetId(key1Id, key2Id);
+            if (key1Id != command.Key1Id)
+            {
+                return BadRequest();
+            }
+            if (key2Id != command.Key2Id)
             {
                 return BadRequest();
             }
@@ -96,36 +101,32 @@ namespace CleanArchitecture.TestApplication.Api.Controllers
 
         /// <summary>
         /// </summary>
-        /// <response code="200">Returns the specified AggregateRootLongDto.</response>
+        /// <response code="200">Returns the specified WithCompositeKeyDto.</response>
         /// <response code="400">One or more validation errors have occurred.</response>
-        /// <response code="404">No AggregateRootLongDto could be found with the provided parameters.</response>
-        [HttpGet("api/aggregate-root-longs/{id}")]
-        [ProducesResponseType(typeof(AggregateRootLongDto), StatusCodes.Status200OK)]
+        /// <response code="404">No WithCompositeKeyDto could be found with the provided parameters.</response>
+        [HttpGet("api/with-composite-key/{key1Id}/{key2Id}")]
+        [ProducesResponseType(typeof(WithCompositeKeyDto), StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult<AggregateRootLongDto>> GetAggregateRootLongById(
-            [FromRoute] long id,
+        public async Task<ActionResult<WithCompositeKeyDto>> GetWithCompositeKeyById(
+            [FromRoute] Guid key1Id,
+            [FromRoute] Guid key2Id,
             CancellationToken cancellationToken = default)
         {
-            var result = await _mediator.Send(new GetAggregateRootLongByIdQuery(id: id), cancellationToken);
+            var result = await _mediator.Send(new GetWithCompositeKeyByIdQuery(key1Id: key1Id, key2Id: key2Id), cancellationToken);
             return result == null ? NotFound() : Ok(result);
         }
 
         /// <summary>
         /// </summary>
-        /// <response code="200">Returns the specified PagedResult&lt;AggregateRootLongDto&gt;.</response>
-        /// <response code="400">One or more validation errors have occurred.</response>
-        [HttpGet("api/aggregate-root-longs")]
-        [ProducesResponseType(typeof(PagedResult<AggregateRootLongDto>), StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        /// <response code="200">Returns the specified List&lt;WithCompositeKeyDto&gt;.</response>
+        [HttpGet("api/with-composite-key")]
+        [ProducesResponseType(typeof(List<WithCompositeKeyDto>), StatusCodes.Status200OK)]
         [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
-        public async Task<ActionResult<PagedResult<AggregateRootLongDto>>> GetAggregateRootLongs(
-            [FromQuery] int pageNo,
-            [FromQuery] int pageSize,
-            CancellationToken cancellationToken = default)
+        public async Task<ActionResult<List<WithCompositeKeyDto>>> GetWithCompositeKeys(CancellationToken cancellationToken = default)
         {
-            var result = await _mediator.Send(new GetAggregateRootLongsQuery(pageNo: pageNo, pageSize: pageSize), cancellationToken);
+            var result = await _mediator.Send(new GetWithCompositeKeysQuery(), cancellationToken);
             return Ok(result);
         }
     }