+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Application/AggregateTestNoIdReturns/UpdateAggregateTestNoIdReturn/UpdateAggregateTestNoIdReturnCommand.cs
@@ -16,7 +16,15 @@ namespace CleanArchitecture.TestApplication.Application.AggregateTestNoIdReturns
             Attribute = attribute;
         }
 
-        public Guid Id { get; set; }
+        public Guid Id { get; private set; }
         public string Attribute { get; set; }
+
+        public void SetId(Guid id)
+        {
+            if (Id == default)
+            {
+                Id = id;
+            }
+        }
     }
 }