+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Application/ImplicitKeyAggrRoots/UpdateImplicitKeyAggrRootImplicitKeyNestedComposition/UpdateImplicitKeyAggrRootImplicitKeyNestedCompositionCommand.cs
@@ -20,7 +20,15 @@ namespace CleanArchitecture.TestApplication.Application.ImplicitKeyAggrRoots.Upd
         }
 
         public Guid ImplicitKeyAggrRootId { get; set; }
-        public Guid Id { get; set; }
+        public Guid Id { get; private set; }
         public string Attribute { get; set; }
+
+        public void SetId(Guid id)
+        {
+            if (Id == default)
+            {
+                Id = id;
+            }
+        }
     }
 }