+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Application/UniqueIndexConstraint/UpdateAggregateWithUniqueConstraintIndexElement/UpdateAggregateWithUniqueConstraintIndexElementCommand.cs
@@ -21,9 +21,17 @@ namespace CleanArchitecture.TestApplication.Application.UniqueIndexConstraint.Up
             CompUniqueFieldB = compUniqueFieldB;
         }
 
-        public Guid Id { get; set; }
+        public Guid Id { get; private set; }
         public string SingleUniqueField { get; set; }
         public string CompUniqueFieldA { get; set; }
         public string CompUniqueFieldB { get; set; }
+
+        public void SetId(Guid id)
+        {
+            if (Id == default)
+            {
+                Id = id;
+            }
+        }
     }
 }