+++ Tests/CleanArchitecture.TestApplication/CleanArchitecture.TestApplication.Application/WithCompositeKeys/UpdateWithCompositeKey/UpdateWithCompositeKeyCommand.cs
@@ -18,7 +18,20 @@ namespace CleanArchitecture.TestApplication.Application.WithCompositeKeys.Update
         }
 
         public string Name { get; set; }
-        public Guid Key1Id { get; set; }
-        public Guid Key2Id { get; set; }
+        public Guid Key1Id { get; private set; }
+        public Guid Key2Id { get; private set; }
+
+        public void SetId(Guid key1Id, Guid key2Id)
+        {
+            if (Key1Id == default)
+            {
+                Key1Id = key1Id;
+            }
+
+            if (Key2Id == default)
+            {
+                Key2Id = key2Id;
+            }
+        }
     }
 }