+++ Modules/Intent.Modules.AspNetCore.Controllers/release-notes.md
@@ -1,3 +1,7 @@
+### Version 5.4.1
+
+- Improvement: Internal model update for better identifying Controller Templates during Software Factory Execution.
+
 ### Version 5.4.0
 
 - Improvement: The heuristic for whether or not to annotate that particular controller methods can return a 404 response has been improved.